+++ src/components/header/Header.tsx
@@ -12,17 +12,10 @@ export const Header: React.FC = () => {
   const router = useRouter();
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [isModalOpen, setModalOpen] = useState(false);
-  const [userName, setUserName] = useState<string | null>(null);
 
   useEffect(() => {
     const accessToken = localStorage.getItem('accessToken');
     setIsLoggedIn(Boolean(accessToken));
-
-    const storedUserInfo = localStorage.getItem('userInfo');
-    if (storedUserInfo) {
-      const { nickname } = JSON.parse(storedUserInfo);
-      setUserName(nickname);
-    }
   }, []);
 
   return (
@@ -38,10 +31,7 @@ export const Header: React.FC = () => {
             <AlarmIcon />
           </button>
           <div className={styles.divider}></div>
-          <div className={styles.userGroup}>
-            <Dropdown />
-            <span className={styles.userName}>{userName}</span>
-          </div>
+          <Dropdown />
         </div>
       ) : (
         // 비 로그인
